add github new project
Using HTML, CSS, and PHP, a contact form is created that can be used on individual websites. The form element is created in HTML, its visual design is improved with CSS, and actions are taken by receiving the form data from PHP, such as sending an email or saving it in the database.